requirements.txt
# DataPulse v5 — Dependencies
# PYTHONPATH=/app must be set for all "from backend.xxx" imports

# ── Web Framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Task Queue ───────────────────────────────────────────────
celery==5.4.0
redis==5.0.8
flower==2.0.1

# ── WebSocket ────────────────────────────────────────────────
websockets==12.0

# ── Auth / Database ──────────────────────────────────────────
supabase==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── AI ───────────────────────────────────────────────────────
anthropic==0.34.2

# ── Data Processing ──────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
xlrd==2.0.1
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.13

# ── Web Scraping ─────────────────────────────────────────────
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.3.0

# ── Google Sheets ────────────────────────────────────────────
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# ── Payments ─────────────────────────────────────────────────
iyzipay==2.0.5

# ── Notifications ────────────────────────────────────────────
httpx==0.27.2

# ── Security ─────────────────────────────────────────────────
python-magic==0.4.27

# ── Utilities ────────────────────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiofiles==24.1.0
structlog==24.4.0
